+++ ui/components/stimulation/StimulationStudioDragAndDropPanel.vue
@@ -50,10 +50,10 @@
               :style="getBorderStyle(pulse)"
             >
               <!-- Only display circular icon when nested loop, icon displays number of loops -->
-              <div v-if="pulse.numRepeats" :class="'div__repeat-label-container'">
-                <div class="div__circle" @dblclick="openRepeatModalForEdit(pulse.numRepeats, idx)">
+              <div v-if="pulse.numInterations" :class="'div__repeat-label-container'">
+                <div class="div__circle" @dblclick="openRepeatModalForEdit(pulse.numInterations, idx)">
                   <span :class="'span__repeat-label'">
-                    {{ pulse.numRepeats }}
+                    {{ pulse.numInterations }}
                   </span>
                 </div>
               </div>
@@ -160,10 +160,10 @@ export default {
     draggable,
     StimulationStudioWaveformSettingModal,
     StimulationStudioInputModal,
-    SmallDropDown
+    SmallDropDown,
   },
   props: {
-    disableEdits: { type: Boolean, default: false }
+    disableEdits: { type: Boolean, default: false },
   },
   data() {
     return {
@@ -187,41 +187,40 @@ export default {
       includeInputUnits: true,
       openRepeatModal: false,
       dblClickPulseNestedIdx: null,
-      hoveredPulse: null
     };
   },
   computed: {
     ...mapState("stimulation", {
-      timeUnit: state => state.protocolEditor.timeUnit,
-      runUntilStopped: state => state.protocolEditor.runUntilStopped,
-      detailedSubprotocols: state => state.protocolEditor.detailedSubprotocols
-    })
+      timeUnit: (state) => state.protocolEditor.timeUnit,
+      runUntilStopped: (state) => state.protocolEditor.runUntilStopped,
+      detailedSubprotocols: (state) => state.protocolEditor.detailedSubprotocols,
+    }),
   },
   watch: {
-    isDragging: function() {
+    isDragging: function () {
       // reset so old position/idx isn't highlighted once moved
       this.onPulseMouseleave();
     },
-    detailedSubprotocols: function() {
+    detailedSubprotocols: function () {
       this.protocolOrder = JSON.parse(
         JSON.stringify(
-          this.detailedSubprotocols.map(protocol =>
+          this.detailedSubprotocols.map((protocol) =>
             protocol.type !== "loop"
               ? {
                   ...protocol,
-                  subprotocols: []
+                  subprotocols: [],
                 }
               : protocol
           )
         )
       );
     },
-    timeUnit: function() {
+    timeUnit: function () {
       this.timeUnitsIdx = this.timeUnitsArray.indexOf(this.timeUnit);
     },
-    runUntilStopped: function() {
+    runUntilStopped: function () {
       this.disableDropdown = !this.runUntilStopped;
-    }
+    },
   },
   methods: {
     ...mapActions("stimulation", ["handleProtocolOrder", "onPulseMouseenter"]),
@@ -353,14 +352,14 @@ export default {
       // create loop object to replace at index in protocol order
       const loopPulse = {
         type: "loop",
-        numRepeats: value,
-        subprotocols: [this.protocolOrder[this.dblClickPulseIdx], this.selectedPulseSettings]
+        numInterations: value,
+        subprotocols: [this.protocolOrder[this.dblClickPulseIdx], this.selectedPulseSettings],
       };
 
       switch (button) {
         case "Save":
           if (this.modalOpenForEdit) {
-            this.protocolOrder[this.dblClickPulseIdx].numRepeats = value;
+            this.protocolOrder[this.dblClickPulseIdx].numInterations = value;
           } else {
             this.protocolOrder.splice(this.dblClickPulseIdx, 1, loopPulse);
           }
@@ -450,12 +449,12 @@ export default {
               frequency: "",
               totalActiveDuration: {
                 duration: "",
-                unit: "milliseconds"
+                unit: "milliseconds",
               },
               numCycles: 0,
               postphaseInterval: "",
               phaseOneDuration: "",
-              phaseOneCharge: ""
+              phaseOneCharge: "",
             };
 
       if (type === "Biphasic")
@@ -463,14 +462,14 @@ export default {
           ...typeSpecificSettings,
           interphaseInterval: "",
           phaseTwoCharge: "",
-          phaseTwoDuration: ""
+          phaseTwoDuration: "",
         };
 
       return {
         type,
         color: randomColor,
         pulseSettings: typeSpecificSettings,
-        subprotocols: []
+        subprotocols: [],
       };
     },
     openRepeatModalForEdit(number, idx) {
@@ -500,8 +499,8 @@ export default {
 
         this.handleProtocolOrder(this.protocolOrder);
       }
-    }
-  }
+    },
+  },
 };
 </script>
 <style scoped>