+++ ui/store/modules/stimulation/actions.js
@@ -2,11 +2,10 @@ import { WellTitle as LabwareDefinition } from "@/js-utils/LabwareCalculations.j
 const twentyFourWellPlateDefinition = new LabwareDefinition(4, 6);
 import { STIM_STATUS, TIME_CONVERSION_TO_MILLIS } from "./enums";
 import {
-  isValidDelayPulse,
-  isValidSinglePulse,
+  areValidPulses,
   convertProtocolCasing,
   _convertObjToCamelCase,
-  _convertObjToSnakeCase
+  _convertObjToSnakeCase,
 } from "@/js-utils/ProtocolValidation";
 
 export default {
@@ -25,19 +24,19 @@ export default {
     const yValues = [0];
     const colorAssignments = [];
     const subprotocols = [];
-    const getLast = array => array[array.length - 1];
+    const getLast = (array) => array[array.length - 1];
     const helper = (setting, type) => {
       let componentsToAdd = [];
       if (type === "Delay") {
         componentsToAdd = {
           x: [setting.duration * TIME_CONVERSION_TO_MILLIS[setting.unit]],
-          y: [0]
+          y: [0],
         };
       } else {
         // Add values for phase 1
         componentsToAdd = {
           x: [setting.phaseOneDuration],
-          y: [setting.phaseOneCharge]
+          y: [setting.phaseOneCharge],
         };
         // If biphasic, handle remaining pulse components
         if (setting.phaseTwoDuration != null) {
@@ -65,7 +64,7 @@ export default {
       yValues.push(0);
     };
 
-    await newSubprotocolOrder.map(async pulse => {
+    await newSubprotocolOrder.map(async (pulse) => {
       if (pulse.type !== "loop") {
         const { color } = pulse;
         let settings = pulse.pulseSettings;
@@ -73,7 +72,7 @@ export default {
 
         settings = {
           type: pulse.type,
-          ...settings
+          ...settings,
         };
 
         subprotocols.push(settings);
@@ -92,15 +91,15 @@ export default {
         const pulseCopy = JSON.parse(JSON.stringify(pulse));
 
         // eslint-disable-next-line  no-unused-vars
-        for (const _ of Array(pulse.numRepeats).fill()) {
-          pulseCopy.subprotocols.map(innerPulse => {
+        for (const _ of Array(pulse.numInterations).fill()) {
+          pulseCopy.subprotocols.map((innerPulse) => {
             const { color } = innerPulse;
             let settings = innerPulse.pulseSettings;
             const startingRepeatIdx = xValues.length - 1;
 
             settings = {
               type: innerPulse.type,
-              ...settings
+              ...settings,
             };
 
             let remainingPulseCycles = innerPulse.type === "Delay" ? 1 : settings.numCycles;
@@ -115,11 +114,11 @@ export default {
           });
         }
 
-        pulseCopy.subprotocols = pulseCopy.subprotocols.map(loopedPulse => {
+        pulseCopy.subprotocols = pulseCopy.subprotocols.map((loopedPulse) => {
           const settings = loopedPulse.pulseSettings;
           return {
             type: loopedPulse.type,
-            ...settings
+            ...settings,
           };
         });
 
@@ -136,7 +135,7 @@ export default {
     commit("setSubprotocols", { subprotocols, newSubprotocolOrder });
     dispatch("handleRestDuration", {
       xValues,
-      yValues
+      yValues,
     });
   },
 
@@ -179,12 +178,12 @@ export default {
   async handleImportProtocol({ dispatch }, file) {
     const reader = new FileReader();
 
-    reader.onload = async function() {
+    reader.onload = async function () {
       const response = JSON.parse(reader.result);
       await dispatch("addImportedProtocol", response);
     };
 
-    reader.onerror = function() {
+    reader.onerror = function () {
       console.log(reader.onerror); // allow-log
     };
 
@@ -194,9 +193,11 @@ export default {
   async handleExportProtocol({ state }) {
     const { protocolAssignments, protocolList } = state;
     // convert to snakecase to be interchangable with MA controller
-    const protocolCopy = [...protocolList].map(protocol =>
-      convertProtocolCasing(protocol, _convertObjToSnakeCase)
-    );
+    const protocolCopy = [];
+    for (const protocol of [...protocolList]) {
+      const convertedProtocol = convertProtocolCasing(protocol.protocol, _convertObjToSnakeCase);
+      protocolCopy.push({ ...protocol, protocol: convertedProtocol });
+    }
 
     const message = { protocols: protocolCopy.slice(1), protocolAssignments: {} };
 
@@ -247,12 +248,9 @@ export default {
       protocol.name = protocol.name.length > 0 ? protocol.name : `protocol_${idx + 1}`;
       // (22/04/2023) For now with mantarray, all protocols will be exported in snake_case, including from stingray
       const convertedProtocol = convertProtocolCasing(protocol, _convertObjToCamelCase);
-      const invalidPulses = convertedProtocol.subprotocols.filter(pulse => {
-        // first check if imported protocols are from stingray or mantarray
-        return !(pulse.type === "Delay" ? isValidDelayPulse(pulse) : isValidSinglePulse(pulse));
-      });
+      const invalidPulses = areValidPulses(convertedProtocol.subprotocols);
 
-      if (invalidPulses.length === 0) {
+      if (invalidPulses) {
         await commit("setEditModeOff");
         // needs to be set to off every iteration because an action elsewhere triggers it on
         const { color, letter } = await getters["getNextProtocol"];
@@ -281,7 +279,7 @@ export default {
           protocolListCopy[idx] = {
             ...protocol,
             label: protocolEditor.name,
-            protocol: protocolEditor
+            protocol: protocolEditor,
           };
       });
       await commit("setProtocolList", protocolListCopy);
@@ -329,7 +327,7 @@ export default {
             protocol_id: letter,
             stimulation_type: stimulationType,
             run_until_stopped: runUntilStopped,
-            subprotocols: convertedSubprotocols
+            subprotocols: convertedSubprotocols,
           };
 
           message.protocols.push(protocolModel);
@@ -387,12 +385,12 @@ export default {
     const { xAxisValues, yAxisValues, xAxisTimeIdx } = state;
 
     if (idx !== xAxisTimeIdx) {
-      const convertedXValues = xAxisValues.map(val => (idx === 1 ? val * 1e-3 : val * 1e3));
+      const convertedXValues = xAxisValues.map((val) => (idx === 1 ? val * 1e-3 : val * 1e3));
       commit("setXAxisTimeIdx", idx);
       if (convertedXValues.length > 0)
         dispatch("handleRestDuration", {
           xValues: convertedXValues,
-          yValues: yAxisValues
+          yValues: yAxisValues,
         });
     }
   },
@@ -406,7 +404,7 @@ export default {
       command: "start_stim_checks",
       well_indices: wellIndices,
       plate_barcode: plateBarcode,
-      stim_barcode: stimBarcode
+      stim_barcode: stimBarcode,
     });
 
     this.state.system.socket.send(wsMessage);
@@ -423,18 +421,18 @@ export default {
         .filter((_, i) => i < idx)
         // reduce to get index
         .reduce((acc, pulse) => {
-          const val = pulse.type === "loop" ? pulse.subprotocols.length * pulse.numRepeats : 1;
+          const val = pulse.type === "loop" ? pulse.subprotocols.length * pulse.numInterations : 1;
           return acc + val;
         }, 0);
 
       // loop through subprotocols x amount of times to highlight every instance in a loop
-      hoveredPulses = [...Array(originalPulse.numRepeats).keys()].map(i => {
+      hoveredPulses = [...Array(originalPulse.numInterations).keys()].map((i) => {
         const numSubprotocols = originalPulse.subprotocols.length;
         const idxToUse = startingIdx + nestedIdx + i * numSubprotocols;
         return {
           idx: idxToUse,
           indices: state.repeatColors[idxToUse][1],
-          color: state.repeatColors[idxToUse][0]
+          color: state.repeatColors[idxToUse][0],
         };
       });
     } else {
@@ -442,7 +440,7 @@ export default {
       const idxToUse = state.protocolEditor.detailedSubprotocols
         .filter((_, i) => i <= idx)
         .reduce((acc, pulse, i) => {
-          let val = pulse.type === "loop" ? pulse.subprotocols.length * pulse.numRepeats : 1;
+          let val = pulse.type === "loop" ? pulse.subprotocols.length * pulse.numInterations : 1;
           if (i === 0) val = val-- < 0 ? 0 : val--;
           return acc + val;
         }, 0);
@@ -451,8 +449,8 @@ export default {
         {
           idx,
           indices: state.repeatColors[idxToUse][1],
-          color: state.repeatColors[idxToUse][0]
-        }
+          color: state.repeatColors[idxToUse][0],
+        },
       ];
     }
     // needs to be array [{}, ... ]
@@ -473,24 +471,24 @@ export default {
         .map(([idx, status]) => {
           return status == "open" ? +idx : undefined;
         })
-        .filter(i => i === 0 || i);
+        .filter((i) => i === 0 || i);
 
       commit("setStimulatorCircuitStatuses", filteredStatuses);
       commit("setStimStatus", STIM_STATUS.CONFIG_CHECK_COMPLETE);
     }
-  }
+  },
 };
 
-const _getConvertedSettings = subprotocols => {
+const _getConvertedSettings = (subprotocols) => {
   const milliToMicro = 1e3;
   const chargeConversion = milliToMicro;
 
-  return subprotocols.map(pulse => {
+  return subprotocols.map((pulse) => {
     let typeSpecificSettings = {};
     if (pulse.type === "loop") {
       typeSpecificSettings = {
-        num_iterations: pulse.numRepeats,
-        subprotocols: _getConvertedSettings(pulse.subprotocols)
+        num_iterations: pulse.numInterations,
+        subprotocols: _getConvertedSettings(pulse.subprotocols),
       };
     } else if (pulse.type === "Delay")
       typeSpecificSettings.duration = pulse.duration * TIME_CONVERSION_TO_MILLIS[pulse.unit] * milliToMicro;
@@ -499,7 +497,7 @@ const _getConvertedSettings = subprotocols => {
         num_cycles: pulse.numCycles,
         postphase_interval: Math.round(pulse.postphaseInterval * milliToMicro), // sent in µs, also needs to be an integer value
         phase_one_duration: pulse.phaseOneDuration * milliToMicro, // sent in µs
-        phase_one_charge: pulse.phaseOneCharge * chargeConversion // sent in mV
+        phase_one_charge: pulse.phaseOneCharge * chargeConversion, // sent in mV
       };
 
     if (pulse.type === "Biphasic")
@@ -507,12 +505,12 @@ const _getConvertedSettings = subprotocols => {
         ...typeSpecificSettings,
         interphase_interval: pulse.interphaseInterval * milliToMicro, // sent in µs
         phase_two_charge: pulse.phaseTwoCharge * chargeConversion, // sent in mV or µA
-        phase_two_duration: pulse.phaseTwoDuration * milliToMicro // sent in µs
+        phase_two_duration: pulse.phaseTwoDuration * milliToMicro, // sent in µs
       };
 
     return {
       type: pulse.type,
-      ...typeSpecificSettings
+      ...typeSpecificSettings,
     };
   });
 };