+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Added:
 - Offline FW updating
 - Protocol validation checks on import in Stimulation Studio
 - Disabled sidebar controls before connection to instrument is made
+- Ability to loop subprotocols in stimulation studio
 
 Fixed:
 ^^^^^^